refactor(renderer): extract web search controller
