+++ src/renderer/src/App.tsx
@@ -13,10 +13,6 @@ import type {
   AppSettings,
   IndexedFileSearchResult,
   BrowserSearchResultGroupSetting,
-  WebSearchBangCustomProviderSetting,
-  WebSearchBangEntry,
-  WebSearchBangOverrideSetting,
-  WebSearchBangUsageSetting,
 } from '../types/electron';
 import ExtensionView from './ExtensionView';
 import ClipboardManager from './ClipboardManager';
@@ -38,6 +34,7 @@ import { useSpeakManager } from './hooks/useSpeakManager';
 import { useWhisperManager } from './hooks/useWhisperManager';
 import { useBrowserSearch } from './hooks/useBrowserSearch';
 import { useBrowserResultsController } from './hooks/useBrowserResultsController';
+import { useWebSearchController } from './hooks/useWebSearchController';
 import { useLauncherCommandModel } from './hooks/useLauncherCommandModel';
 import { useLauncherInlineArguments } from './hooks/useLauncherInlineArguments';
 import { useLauncherActionModel } from './hooks/useLauncherActionModel';
@@ -74,11 +71,7 @@ import AiChatView from './views/AiChatView';
 import CursorPromptView from './views/CursorPromptView';
 import AppUninstallView from './views/AppUninstallView';
 import BrowserResultsView from './views/BrowserResultsView';
-import WebSearchView, {
-  type WebSearchBangPromptState,
-  type WebSearchCustomBangPromptState,
-  type WebSearchViewSection,
-} from './views/WebSearchView';
+import WebSearchView from './views/WebSearchView';
 import LauncherMainView from './views/LauncherMainView';
 import HiddenExtensionRunners from './components/HiddenExtensionRunners';
 import DetachedOverlayRunners from './components/DetachedOverlayRunners';
@@ -116,31 +109,9 @@ import {
   getQuickLinkIdFromCommandId,
 } from './utils/launcher-misc';
 import {
-  type BangParseState,
-  type SearchBangDefinition,
-  type WebSearchResult,
-  WEB_SEARCH_ACTIVE_BANG_SUGGESTION_LIMIT,
-  WEB_SEARCH_BANG_USE_COUNTS_KEY,
-  WEB_SEARCH_COMMAND_ID,
-  WEB_SEARCH_INITIAL_VISIBLE_RESULTS,
-  WEB_SEARCH_RECENT_BANG_LIMIT,
   WEB_SEARCH_ROOT_BANG_PREFIX,
-  WEB_SEARCH_SUGGEST_DEBOUNCE_MS,
-  WEB_SEARCH_VISIBLE_RESULTS_INCREMENT,
-  SEARCH_BANGS,
   buildBangSearchUrl,
-  createUpdatedBangUsage,
-  formatWebSearchBangAliases,
-  formatWebSearchBangAliasSummary,
-  getBangUsageScore,
-  getFaviconUrlForHost,
   getSearchBangByKeyFromList,
-  getSortedSearchBangs,
-  getWebSearchBangSection,
-  getWebSearchBangSectionTitleKey,
-  normalizeBangDefinition,
-  normalizeWebSearchBangAliasList,
-  parseSearchBangFromList,
   parseSearchBangState,
 } from './utils/web-search-bangs';
 
@@ -174,18 +145,6 @@ const App: React.FC = () => {
   const [browserSearchResultGroups, setBrowserSearchResultGroups] = useState<BrowserSearchResultGroupSetting[]>(
     DEFAULT_BROWSER_SEARCH_RESULT_GROUPS
   );
-  const [webSearchQuery, setWebSearchQuery] = useState<string | null>(null);
-  const [webSearchSelectedIndex, setWebSearchSelectedIndex] = useState(0);
-  const [rootWebSearchSuggestions, setRootWebSearchSuggestions] = useState<string[]>([]);
-  const [webSearchSuggestions, setWebSearchSuggestions] = useState<string[]>([]);
-  const [webSearchDefaultBangKey, setWebSearchDefaultBangKey] = useState('g');
-  const [webSearchSuggestionLimit, setWebSearchSuggestionLimit] = useState(3);
-  const [webSearchBangUsage, setWebSearchBangUsage] = useState<Record<string, WebSearchBangUsageSetting>>({});
-  const [webSearchBangCatalog, setWebSearchBangCatalog] = useState<SearchBangDefinition[]>([]);
-  const [webSearchBangOverrides, setWebSearchBangOverrides] = useState<WebSearchBangOverrideSetting[]>([]);
-  const [webSearchDisabledBangKeys, setWebSearchDisabledBangKeys] = useState<string[]>([]);
-  const [webSearchBangCustomProviders, setWebSearchBangCustomProviders] = useState<WebSearchBangCustomProviderSetting[]>([]);
-  const [webSearchShowHiddenBangs, setWebSearchShowHiddenBangs] = useState(false);
   const [launcherFileResults, setLauncherFileResults] = useState<IndexedFileSearchResult[]>([]);
   const [disableFileSearchResults, setDisableFileSearchResults] = useState(false);
   const [launcherViewMode, setLauncherViewMode] = useState<'expanded' | 'compact'>('expanded');
@@ -241,9 +200,6 @@ const App: React.FC = () => {
   const [quickLinkEditId, setQuickLinkEditId] = useState<string | null>(null);
   const [quickLinkDynamicPrompt, setQuickLinkDynamicPrompt] =
     useState<QuickLinkDynamicPromptState | null>(null);
-  const [webSearchBangPrompt, setWebSearchBangPrompt] = useState<WebSearchBangPromptState | null>(null);
-  const [webSearchCustomBangPrompt, setWebSearchCustomBangPrompt] = useState<WebSearchCustomBangPromptState | null>(null);
-  const [webSearchVisibleResultCount, setWebSearchVisibleResultCount] = useState(WEB_SEARCH_INITIAL_VISIBLE_RESULTS);
   const {
     menuBarExtensions,
     backgroundNoViewRuns, setBackgroundNoViewRuns,
@@ -298,13 +254,13 @@ const App: React.FC = () => {
     launcherInputRef: inputRef,
     t,
   });
-  const webSearchInputRef = useRef<HTMLInputElement>(null);
+  const submitBrowserSearchRef = useRef<
+    (query: string, options?: { focusExistingTab?: boolean }) => void | Promise<boolean>
+  >(() => false);
   const isLauncherModeActiveRef = useRef(false);
   const fileSearchRequestSeqRef = useRef(0);
   const commandsRef = useRef<CommandInfo[]>([]);
   const lastCommandsFetchAtRef = useRef(0);
-  const lastRecordedRootBangUseRef = useRef<string | null>(null);
-  const lastRecordedWebBangUseRef = useRef<string | null>(null);
   const executingCommandRef = useRef(false);
   const showActionsRef = useRef(false);
   const showAppUninstallRef = useRef<string | null>(null);
@@ -368,7 +324,6 @@ const App: React.FC = () => {
   const actionsOverlayRef = useRef<HTMLDivElement>(null);
   const contextMenuRef = useRef<HTMLDivElement>(null);
   const quickLinkDynamicInputRef = useRef<HTMLInputElement>(null);
-  const webSearchBangInputRef = useRef<HTMLInputElement>(null);
   const windowPresetCommandQueueRef = useRef<Promise<void>>(Promise.resolve());
   const lastWindowHiddenAtRef = useRef<number>(0);
   const directLaunchExpansionGuardUntilRef = useRef<number>(0);
@@ -381,44 +336,80 @@ const App: React.FC = () => {
   extensionViewRef.current = extensionView;
   pinnedCommandsRef.current = pinnedCommands;
   pinnedFilesRef.current = pinnedFiles;
-  const effectiveSearchBangs = useMemo(() => {
-    const byKey = new Map<string, SearchBangDefinition>();
-    const disabled = new Set(webSearchDisabledBangKeys);
-    for (const entry of webSearchBangCatalog) {
-      const normalized = normalizeBangDefinition(entry);
-      byKey.set(normalized.key, { ...normalized, disabled: disabled.has(normalized.key) });
-    }
-    for (const entry of SEARCH_BANGS) {
-      const normalized = normalizeBangDefinition(entry);
-      byKey.set(normalized.key, { ...normalized, disabled: disabled.has(normalized.key) });
-    }
-    for (const entry of webSearchBangCustomProviders) {
-      const normalized = normalizeBangDefinition({
-        key: entry.key,
-        aliases: entry.aliases,
-        name: entry.name,
-        host: entry.host,
-        template: entry.template,
-        category: 'Custom',
-        source: 'custom',
-      });
-      byKey.set(normalized.key, { ...normalized, disabled: disabled.has(normalized.key) });
-    }
-    const overrides = new Map(webSearchBangOverrides.map((override) => [override.key, override]));
-    for (const [key, override] of overrides) {
-      const current = byKey.get(key);
-      if (!current) continue;
-      byKey.set(key, {
-        ...current,
-        aliases: override.aliases.filter((alias) => alias !== key),
-      });
-    }
-    return Array.from(byKey.values());
-  }, [webSearchBangCatalog, webSearchBangOverrides, webSearchBangCustomProviders, webSearchDisabledBangKeys]);
-  const enabledSearchBangs = useMemo(
-    () => effectiveSearchBangs.filter((bang) => !bang.disabled),
-    [effectiveSearchBangs]
-  );
+
+  const expandLauncherForDirectLaunch = useCallback(() => {
+    directLaunchExpansionGuardUntilRef.current = Date.now() + DIRECT_LAUNCH_EXPANSION_GUARD_MS;
+    setIsCompactCollapsed(false);
+    void window.electron.resizeLauncherWindow(true);
+
+    // Extension/script direct launches are dispatched with executeJavaScript(),
+    // which can beat the async window-shown IPC reset. Retry briefly so the
+    // direct launch remains expanded regardless of delivery order.
+    [0, 80, 180].forEach((delayMs) => {
+      window.setTimeout(() => {
+        if (Date.now() > directLaunchExpansionGuardUntilRef.current) return;
+        setIsCompactCollapsed(false);
+        void window.electron.resizeLauncherWindow(true);
+      }, delayMs);
+    });
+  }, []);
+
+  const {
+    webSearchQuery,
+    setWebSearchQuery,
+    webSearchSelectedIndex,
+    setWebSearchSelectedIndex,
+
+    webSearchInputRef,
+    webSearchBangInputRef,
+
+    webSearchDefaultBangKey,
+    webSearchSuggestionLimit,
+    webSearchBangUsage,
+    webSearchShowHiddenBangs,
+
+    effectiveSearchBangs,
+    enabledSearchBangs,
+    rootBangState,
+    rootWebSearchSuggestions,
+
+    webSearchResults,
+    visibleWebSearchSections,
+    selectedWebSearchResult,
+    isWebSearchBangManager,
+    activeWebSearchBang,
+
+    webSearchBangPrompt,
+    setWebSearchBangPrompt,
+    webSearchCustomBangPrompt,
+    setWebSearchCustomBangPrompt,
+
+    openWebSearchMode,
+    closeWebSearch,
+    activateWebSearchResult,
+    loadMoreWebSearchResults,
+
+    openWebSearchBangPrompt,
+    saveWebSearchBangAliases,
+    toggleWebSearchBangDisabled,
+    toggleWebSearchShowHidden,
+
+    openWebSearchCustomBangPrompt,
+    closeWebSearchCustomBangPrompt,
+    saveWebSearchCustomBang,
+
+    hydrateWebSearchSettings,
+  } = useWebSearchController({
+    launcherInputRef: inputRef,
+    expandLauncherForDirectLaunch,
+    submitBrowserSearchRef,
+    setLauncherSearchQuery: setSearchQuery,
+    setLauncherSelectedIndex: setSelectedIndex,
+    rootSearchQuery: searchQuery,
+    aiMode,
+    t,
+  });
+
   // Configurable timeout (ms) before the launcher resets to root search after
   // it has been hidden. Synced from settings.popToRootSearchTimeoutSeconds.
   // 0 = reset immediately on every reopen.
@@ -437,23 +428,6 @@ const App: React.FC = () => {
     showSchedule ||
     showAppUninstall
   );
-
-  const expandLauncherForDirectLaunch = useCallback(() => {
-    directLaunchExpansionGuardUntilRef.current = Date.now() + DIRECT_LAUNCH_EXPANSION_GUARD_MS;
-    setIsCompactCollapsed(false);
-    void window.electron.resizeLauncherWindow(true);
-
-    // Extension/script direct launches are dispatched with executeJavaScript(),
-    // which can beat the async window-shown IPC reset. Retry briefly so the
-    // direct launch remains expanded regardless of delivery order.
-    [0, 80, 180].forEach((delayMs) => {
-      window.setTimeout(() => {
-        if (Date.now() > directLaunchExpansionGuardUntilRef.current) return;
-        setIsCompactCollapsed(false);
-        void window.electron.resizeLauncherWindow(true);
-      }, delayMs);
-    });
-  }, []);
 
 
   const cursorPromptPortalTarget = useDetachedPortalWindow(showCursorPrompt, {
@@ -553,13 +527,7 @@ const App: React.FC = () => {
       );
       setLauncherShortcut(settings.globalShortcut || 'Alt+Space');
       setBrowserSearchResultGroups(normalizeBrowserSearchResultGroups(settings.browserSearch?.resultGroups));
-      setWebSearchDefaultBangKey(String(settings.browserSearch?.webSearchDefaultBangKey || 'g'));
-      setWebSearchSuggestionLimit(Math.max(0, Math.min(8, Math.floor(Number(settings.browserSearch?.webSearchSuggestionLimit ?? 3)))));
-      setWebSearchBangOverrides(Array.isArray(settings.browserSearch?.webSearchBangOverrides) ? settings.browserSearch.webSearchBangOverrides : []);
-      setWebSearchBangUsage(settings.browserSearch?.webSearchBangUsage && typeof settings.browserSearch.webSearchBangUsage === 'object' ? settings.browserSearch.webSearchBangUsage : {});
-      setWebSearchDisabledBangKeys(Array.isArray(settings.browserSearch?.webSearchDisabledBangKeys) ? settings.browserSearch.webSearchDisabledBangKeys : []);
-      setWebSearchBangCustomProviders(Array.isArray(settings.browserSearch?.webSearchBangCustomProviders) ? settings.browserSearch.webSearchBangCustomProviders : []);
-      setWebSearchShowHiddenBangs(Boolean(settings.browserSearch?.webSearchShowHiddenBangs));
+      hydrateWebSearchSettings(settings);
       const speakToggleHotkey = settings.commandHotkeys?.['system-supercmd-whisper-speak-toggle'] ?? '';
       setWhisperSpeakToggleLabel(formatShortcutLabel(speakToggleHotkey));
       setConfiguredEdgeTtsVoice(String(settings.ai?.edgeTtsVoice || 'en-US-EricNeural'));
@@ -627,7 +595,7 @@ const App: React.FC = () => {
       setShowOnboarding(false);
       setOnboardingRequiresShortcutFix(false);
     }
-  }, []);
+  }, [hydrateWebSearchSettings]);
 
   const fetchCommands = useCallback(async (options?: { showLoading?: boolean }) => {
     const shouldShowLoading = options?.showLoading ?? commandsRef.current.length === 0;
@@ -789,54 +757,6 @@ const App: React.FC = () => {
   }, []);
 
   useEffect(() => {
-    try {
-      const parsed = JSON.parse(localStorage.getItem(WEB_SEARCH_BANG_USE_COUNTS_KEY) || '{}');
-      if (!parsed || typeof parsed !== 'object' || Array.isArray(parsed)) return;
-      setWebSearchBangUsage((current) => {
-        if (Object.keys(current).length > 0) return current;
-        return Object.entries(parsed).reduce((acc, [key, value]) => {
-          const normalizedKey = String(key || '').trim().toLowerCase();
-          const count = Math.floor(Number(value));
-          if (normalizedKey && Number.isFinite(count) && count > 0) {
-            acc[normalizedKey] = { useCount: count, lastUsedAt: Date.now(), frecencyScore: count };
-          }
-          return acc;
-        }, {} as Record<string, WebSearchBangUsageSetting>);
-      });
-    } catch {}
-  }, []);
-
-  useEffect(() => {
-    let cancelled = false;
-    window.electron.webSearchListBangs?.()
-      .then((entries: WebSearchBangEntry[]) => {
-        if (cancelled || !Array.isArray(entries)) return;
-        const next = entries
-          .map((entry): SearchBangDefinition | null => {
-            const key = String(entry?.key || '').trim().toLowerCase().replace(/^!+/, '');
-            if (!key) return null;
-            return {
-              key,
-              aliases: Array.isArray(entry.aliases) ? entry.aliases : [],
-              name: String(entry.name || key),
-              host: String(entry.host || 'duckduckgo.com'),
-              category: entry.category,
-              subcategory: entry.subcategory,
-              template: String(entry.urlTemplate || 'https://duckduckgo.com/?q=!{bang}%20{query}'),
-              source: entry.source || 'duckduckgo',
-              rankHint: entry.rankHint,
-            };
-          })
-          .filter((entry): entry is SearchBangDefinition => Boolean(entry));
-        setWebSearchBangCatalog(next);
-      })
-      .catch(() => {});
-    return () => {
-      cancelled = true;
-    };
-  }, []);
-
-  useEffect(() => {
     const cleanup = window.electron.onSettingsUpdated?.((settings: AppSettings) => {
       // Settings broadcasts fire for in-app saves AND for external sync
       // changes (cloud watcher → reload → broadcast). Re-hydrate localStorage
@@ -861,20 +781,14 @@ const App: React.FC = () => {
       );
       setLauncherShortcut(settings.globalShortcut || 'Alt+Space');
       setBrowserSearchResultGroups(normalizeBrowserSearchResultGroups(settings.browserSearch?.resultGroups));
-      setWebSearchDefaultBangKey(String(settings.browserSearch?.webSearchDefaultBangKey || 'g'));
-      setWebSearchSuggestionLimit(Math.max(0, Math.min(8, Math.floor(Number(settings.browserSearch?.webSearchSuggestionLimit ?? 3)))));
-      setWebSearchBangOverrides(Array.isArray(settings.browserSearch?.webSearchBangOverrides) ? settings.browserSearch.webSearchBangOverrides : []);
-      setWebSearchBangUsage(settings.browserSearch?.webSearchBangUsage && typeof settings.browserSearch.webSearchBangUsage === 'object' ? settings.browserSearch.webSearchBangUsage : {});
-      setWebSearchDisabledBangKeys(Array.isArray(settings.browserSearch?.webSearchDisabledBangKeys) ? settings.browserSearch.webSearchDisabledBangKeys : []);
-      setWebSearchBangCustomProviders(Array.isArray(settings.browserSearch?.webSearchBangCustomProviders) ? settings.browserSearch.webSearchBangCustomProviders : []);
-      setWebSearchShowHiddenBangs(Boolean(settings.browserSearch?.webSearchShowHiddenBangs));
+      hydrateWebSearchSettings(settings);
       setDisableFileSearchResults(Boolean(settings.disableFileSearchResults));
       setNavigationStyle(settings.navigationStyle === 'macos' ? 'macos' : 'vim');
       const popToRootSeconds = Number(settings.popToRootSearchTimeoutSeconds);
       popToRootTimeoutMsRef.current = (Number.isFinite(popToRootSeconds) ? Math.max(0, popToRootSeconds) : DEFAULT_POP_TO_ROOT_TIMEOUT_SECONDS) * 1000;
     });
     return cleanup;
-  }, []);
+  }, [hydrateWebSearchSettings]);
 
   // Onboarding is intentionally always shown in dark mode for consistent
   // contrast and readability, independent of the user's regular theme.
@@ -1518,47 +1432,15 @@ const App: React.FC = () => {
     if (aiMode) return null;
     if (browserSearchSkipAutoComplete) return null;
     if (!searchQuery) return null;
-    const parsedBang = parseSearchBangState(searchQuery, enabledSearchBangs);
-    if (parsedBang.mode !== 'none') return null;
+    if (rootBangState.mode !== 'none') return null;
     const completion = browserSearch.getCompletion(searchQuery, browserSearchResultGroups);
     if (!completion) return null;
     if (completion.completion === searchQuery) return null;
     if (!completion.completion.toLowerCase().startsWith(searchQuery.toLowerCase())) return null;
     return completion;
-  }, [browserSearch, browserSearchResultGroups, searchQuery, browserSearchSkipAutoComplete, aiMode, enabledSearchBangs]);
+  }, [browserSearch, browserSearchResultGroups, searchQuery, browserSearchSkipAutoComplete, aiMode, rootBangState]);
 
   const launcherInputValue = browserSearchAutoComplete?.completion ?? searchQuery;
-
-  const rootBangState = useMemo<BangParseState>(() => {
-    if (aiMode) return { mode: 'none' };
-    return parseSearchBangState(searchQuery, enabledSearchBangs);
-  }, [aiMode, searchQuery, enabledSearchBangs]);
-
-  useEffect(() => {
-    if (aiMode) return;
-    const query = (rootBangState.mode === 'active' ? rootBangState.query : searchQuery).trim();
-    if (!query || (rootBangState.mode !== 'active' && webSearchSuggestionLimit <= 0)) {
-      setRootWebSearchSuggestions([]);
-      return;
-    }
-    const provider = rootBangState.mode === 'active' ? rootBangState.bang : undefined;
-    const limit = rootBangState.mode === 'active' ? WEB_SEARCH_ACTIVE_BANG_SUGGESTION_LIMIT : webSearchSuggestionLimit;
-    let cancelled = false;
-    const timer = window.setTimeout(() => {
-      window.electron.browserSearchSuggestMany(query, limit, provider ? { key: provider.key, host: provider.host, name: provider.name } : undefined)
-        .then((suggestions) => {
-          if (cancelled) return;
-          setRootWebSearchSuggestions(Array.isArray(suggestions) ? suggestions.slice(0, limit) : []);
-        })
-        .catch(() => {
-          if (!cancelled) setRootWebSearchSuggestions([]);
-        });
-    }, WEB_SEARCH_SUGGEST_DEBOUNCE_MS);
-    return () => {
-      cancelled = true;
-      window.clearTimeout(timer);
-    };
-  }, [aiMode, searchQuery, webSearchSuggestionLimit, rootBangState]);
 
   const {
     calcResult,
@@ -1752,28 +1634,6 @@ const App: React.FC = () => {
     }
   }, [searchQuery, browserSearchSkipAutoComplete]);
 
-  const recordWebSearchBangUse = useCallback((bangKey: string) => {
-    const normalizedKey = String(bangKey || '').trim().toLowerCase();
-    if (!normalizedKey) return;
-    setWebSearchBangUsage((current) => {
-      const next = { ...current, [normalizedKey]: createUpdatedBangUsage(current[normalizedKey]) };
-      try {
-        localStorage.setItem(WEB_SEARCH_BANG_USE_COUNTS_KEY, JSON.stringify(Object.fromEntries(
-          Object.entries(next).map(([key, value]) => [key, value.useCount])
-        )));
-      } catch {}
-      window.electron.getSettings()
-        .then((settings) => window.electron.saveSettings({
-          browserSearch: {
-            ...settings.browserSearch,
-            webSearchBangUsage: next,
-          },
-        }))
-        .catch(() => {});
-      return next;
-    });
-  }, []);
-
   const submitBrowserSearch = useCallback(
     async (input: string, options?: { focusExistingTab?: boolean }) => {
       const trimmed = input.trim();
@@ -1804,461 +1664,10 @@ const App: React.FC = () => {
       }
       return ok;
     },
-    [browserSearch, recordWebSearchBangUse, webSearchDefaultBangKey, effectiveSearchBangs, enabledSearchBangs]
+    [browserSearch, webSearchDefaultBangKey, effectiveSearchBangs, enabledSearchBangs]
   );
 
-  useEffect(() => {
-    if (rootBangState.mode !== 'active') {
-      lastRecordedRootBangUseRef.current = null;
-      return;
-    }
-    const key = rootBangState.bang.key;
-    if (lastRecordedRootBangUseRef.current === key) return;
-    lastRecordedRootBangUseRef.current = key;
-    recordWebSearchBangUse(key);
-  }, [recordWebSearchBangUse, rootBangState]);
-
-  const closeWebSearch = useCallback(() => {
-    setWebSearchQuery(null);
-    setWebSearchSelectedIndex(0);
-    setWebSearchSuggestions([]);
-    setTimeout(() => inputRef.current?.focus(), 50);
-  }, []);
-
-  const openWebSearchMode = useCallback((initialQuery = '') => {
-    expandLauncherForDirectLaunch();
-    setWebSearchQuery(initialQuery);
-    setWebSearchSelectedIndex(0);
-    window.setTimeout(() => webSearchInputRef.current?.focus(), 0);
-  }, [expandLauncherForDirectLaunch]);
-
-  const webSearchBangState = useMemo<BangParseState>(() => {
-    if (webSearchQuery === null) return { mode: 'none' };
-    return parseSearchBangState(webSearchQuery, enabledSearchBangs);
-  }, [enabledSearchBangs, webSearchQuery]);
-
-  useEffect(() => {
-    if (webSearchBangState.mode !== 'active') {
-      lastRecordedWebBangUseRef.current = null;
-      return;
-    }
-    const key = webSearchBangState.bang.key;
-    if (lastRecordedWebBangUseRef.current === key) return;
-    lastRecordedWebBangUseRef.current = key;
-    recordWebSearchBangUse(key);
-  }, [recordWebSearchBangUse, webSearchBangState]);
-
-  useEffect(() => {
-    if (webSearchQuery === null) return;
-    const raw = String(webSearchQuery || '').trim();
-    const searchSubject = webSearchBangState.mode === 'active' ? webSearchBangState.query.trim() : raw;
-    const shouldFetch = Boolean(searchSubject) && (webSearchBangState.mode === 'active' || webSearchBangState.mode === 'none');
-    if (!shouldFetch || (webSearchBangState.mode !== 'active' && webSearchSuggestionLimit <= 0)) {
-      setWebSearchSuggestions([]);
-      return;
-    }
-    const provider = webSearchBangState.mode === 'active'
-      ? webSearchBangState.bang
-      : getSearchBangByKeyFromList(webSearchDefaultBangKey, effectiveSearchBangs);
-    const limit = webSearchBangState.mode === 'active'
-      ? WEB_SEARCH_ACTIVE_BANG_SUGGESTION_LIMIT
-      : webSearchSuggestionLimit;
-    let cancelled = false;
-    const timer = window.setTimeout(() => {
-      window.electron.browserSearchSuggestMany(searchSubject, limit, { key: provider.key, host: provider.host, name: provider.name })
-        .then((suggestions) => {
-          if (cancelled) return;
-          setWebSearchSuggestions(Array.isArray(suggestions) ? suggestions.slice(0, limit) : []);
-        })
-        .catch(() => {
-          if (!cancelled) setWebSearchSuggestions([]);
-        });
-    }, WEB_SEARCH_SUGGEST_DEBOUNCE_MS);
-    return () => {
-      cancelled = true;
-      window.clearTimeout(timer);
-    };
-  }, [effectiveSearchBangs, webSearchBangState, webSearchDefaultBangKey, webSearchQuery, webSearchSuggestionLimit]);
-
-  const webSearchResults = useMemo<WebSearchResult[]>(() => {
-    if (webSearchQuery === null) return [];
-    const raw = String(webSearchQuery || '').trim();
-    const results: WebSearchResult[] = [];
-    if (raw && (webSearchBangState.mode === 'none' || webSearchBangState.mode === 'active')) {
-      const activeBang = webSearchBangState.mode === 'active' ? webSearchBangState.bang : null;
-      const defaultBang = getSearchBangByKeyFromList(webSearchDefaultBangKey, effectiveSearchBangs);
-      const provider = activeBang || defaultBang;
-      const searchSubject = (webSearchBangState.mode === 'active' ? webSearchBangState.query : raw).trim();
-      if (!searchSubject) return [];
-      results.push({
-        id: 'web-search-mode:direct',
-        kind: 'search',
-        section: 'search',
-        title: activeBang
-          ? t('launcher.browserSearch.searchProviderFor', { provider: provider.name, query: searchSubject })
-          : t('launcher.browserSearch.searchFor', { query: searchSubject }),
-        subtitle: activeBang
-          ? t('launcher.browserSearch.bangSubtitle', { bang: activeBang.key })
-          : t('launcher.browserSearch.defaultSearch'),
-        query: activeBang ? `${searchSubject} !${activeBang.key}` : searchSubject,
-        bangKey: activeBang?.key,
-        bang: activeBang || undefined,
-        faviconUrl: getFaviconUrlForHost(provider.host),
-      });
-      for (const suggestion of webSearchSuggestions) {
-        const normalized = String(suggestion || '').trim();
-        if (!normalized || normalized.toLowerCase() === searchSubject.toLowerCase()) continue;
-        results.push({
-          id: `web-search-mode:suggestion:${normalized}`,
-          kind: 'suggestion',
-          section: 'search',
-          title: normalized,
-          subtitle: activeBang
-            ? t('launcher.browserSearch.bangSubtitle', { bang: activeBang.key })
-            : t('launcher.browserSearch.defaultSearch'),
-          query: activeBang ? `${normalized} !${activeBang.key}` : normalized,
-          bangKey: activeBang?.key,
-          bang: activeBang || undefined,
-          faviconUrl: getFaviconUrlForHost(provider.host),
-        });
-      }
-      return results;
-    }
-
-    const parsed = parseSearchBangFromList(raw, enabledSearchBangs);
-    const bangFilter = parsed.activeBangPrefix !== null
-      ? parsed.activeBangPrefix
-      : raw.replace(/^!+/, '');
-    const sectionFilter = parsed.bang && !parsed.query ? '' : bangFilter;
-    const candidateSource = webSearchShowHiddenBangs
-      ? effectiveSearchBangs.filter((bang) => bang.disabled)
-      : enabledSearchBangs;
-    const sorted = parsed.bang && !parsed.query
-      ? [parsed.bang, ...getSortedSearchBangs(candidateSource, null, webSearchBangUsage).filter((bang) => bang.key !== parsed.bang?.key)]
-      : getSortedSearchBangs(candidateSource, bangFilter, webSearchBangUsage);
-    const recentKeys = new Set<string>();
-    const recentBangs = !sectionFilter && !webSearchShowHiddenBangs
-      ? [...sorted]
-          .filter((bang) => getBangUsageScore(webSearchBangUsage[bang.key]) > 0)
-          .sort((a, b) => getBangUsageScore(webSearchBangUsage[b.key]) - getBangUsageScore(webSearchBangUsage[a.key]))
-          .slice(0, WEB_SEARCH_RECENT_BANG_LIMIT)
-      : [];
-    for (const bang of recentBangs) recentKeys.add(bang.key);
-    const matchingBangs = webSearchShowHiddenBangs
-      ? sorted
-      : sectionFilter
-        ? sorted
-        : [...recentBangs, ...sorted.filter((bang) => !recentKeys.has(bang.key))];
-    for (const bang of matchingBangs) {
-      const defaultAliases = [bang.key, ...(bang.aliases || [])];
-      const aliasSummary = formatWebSearchBangAliasSummary(defaultAliases);
-      const baseSubtitle = [bang.category, bang.subcategory, bang.host].filter(Boolean).join(' - ') || bang.host;
-      results.push({
-        id: `web-search-result:bang:${bang.key}`,
-        kind: 'bang',
-        section: webSearchShowHiddenBangs
-          ? 'hidden'
-          : recentKeys.has(bang.key) && !sectionFilter
-            ? 'recent'
-            : getWebSearchBangSection(bang, sectionFilter, webSearchBangUsage),
-        title: `!${bang.key} ${bang.name}`,
-        subtitle: aliasSummary ? `${baseSubtitle} - ${aliasSummary}` : baseSubtitle,
-        query: `!${bang.key} `,
-        bangKey: bang.key,
-        defaultAliases,
-        customAliases: webSearchBangOverrides.find((override) => override.key === bang.key)?.aliases,
-        isCustom: webSearchBangOverrides.some((override) => override.key === bang.key),
-        isDisabled: Boolean(bang.disabled),
-        bang,
-        faviconUrl: getFaviconUrlForHost(bang.host),
-      });
-    }
-    return results;
-  }, [effectiveSearchBangs, enabledSearchBangs, t, webSearchBangOverrides, webSearchBangState, webSearchBangUsage, webSearchDefaultBangKey, webSearchQuery, webSearchShowHiddenBangs, webSearchSuggestions]);
-
-  const visibleWebSearchResults = useMemo(
-    () => webSearchResults.slice(0, Math.min(webSearchVisibleResultCount, webSearchResults.length)),
-    [webSearchResults, webSearchVisibleResultCount]
-  );
-
-  const visibleWebSearchSections = useMemo(() => {
-    const sections: WebSearchViewSection[] = [];
-    const indexByKey = new Map<WebSearchResult['section'], number>();
-    visibleWebSearchResults.forEach((result, flatIndex) => {
-      const sectionIndex = indexByKey.get(result.section);
-      if (sectionIndex === undefined) {
-        indexByKey.set(result.section, sections.length);
-        sections.push({
-          key: result.section,
-          titleKey: getWebSearchBangSectionTitleKey(result.section),
-          items: [result],
-          startIndex: flatIndex,
-        });
-        return;
-      }
-      sections[sectionIndex].items.push(result);
-    });
-    return sections;
-  }, [visibleWebSearchResults]);
-
-  const selectedWebSearchResult = webSearchResults[webSearchSelectedIndex] || null;
-
-  useEffect(() => {
-    setWebSearchSelectedIndex(0);
-    setWebSearchVisibleResultCount(WEB_SEARCH_INITIAL_VISIBLE_RESULTS);
-  }, [webSearchQuery]);
-
-  useEffect(() => {
-    if (webSearchQuery === null) return;
-    setWebSearchSelectedIndex((index) => Math.min(index, Math.max(0, webSearchResults.length - 1)));
-  }, [webSearchQuery, webSearchResults.length]);
-
-  useEffect(() => {
-    if (webSearchQuery === null) return;
-    if (webSearchSelectedIndex < webSearchVisibleResultCount - 12) return;
-    setWebSearchVisibleResultCount((count) =>
-      Math.min(webSearchResults.length, count + WEB_SEARCH_VISIBLE_RESULTS_INCREMENT)
-    );
-  }, [webSearchQuery, webSearchResults.length, webSearchSelectedIndex, webSearchVisibleResultCount]);
-
-  useEffect(() => {
-    if (webSearchQuery === null) return;
-    window.setTimeout(() => webSearchInputRef.current?.focus(), 0);
-  }, [webSearchQuery]);
-
-  useEffect(() => {
-    if (!webSearchBangPrompt) return;
-    const timer = window.setTimeout(() => {
-      webSearchBangInputRef.current?.focus();
-      webSearchBangInputRef.current?.select();
-    }, 0);
-    return () => window.clearTimeout(timer);
-  }, [webSearchBangPrompt?.result.id]);
-
-  const activateWebSearchResult = useCallback(async (result: WebSearchResult | null) => {
-    if (!result) return;
-    if (result.kind === 'bang') {
-      setWebSearchQuery(null);
-      setWebSearchSelectedIndex(0);
-      setSearchQuery(result.query);
-      setSelectedIndex(0);
-      window.setTimeout(() => inputRef.current?.focus(), 0);
-      return;
-    }
-    await submitBrowserSearch(result.query);
-  }, [submitBrowserSearch]);
-
-  const openWebSearchBangPrompt = useCallback((result: WebSearchResult | null) => {
-    if (!result || result.kind !== 'bang' || !result.bangKey) return;
-    setWebSearchBangPrompt({
-      result,
-      value: formatWebSearchBangAliases(result.customAliases || result.defaultAliases || [result.bangKey]),
-    });
-  }, []);
-
-  const closeWebSearchBangPrompt = useCallback(() => {
-    setWebSearchBangPrompt(null);
-    window.setTimeout(() => webSearchInputRef.current?.focus(), 0);
-  }, []);
-
-  const saveWebSearchBangAliases = useCallback(async () => {
-    if (!webSearchBangPrompt?.result.bangKey) return;
-    const key = webSearchBangPrompt.result.bangKey;
-    const aliases = normalizeWebSearchBangAliasList(webSearchBangPrompt.value);
-    const defaultAliases = normalizeWebSearchBangAliasList(
-      formatWebSearchBangAliases(webSearchBangPrompt.result.defaultAliases || [key])
-    );
-    const changed = aliases.join(',') !== defaultAliases.join(',');
-    try {
-      const currentSettings = await window.electron.getSettings();
-      const browserSearchSettings = currentSettings.browserSearch;
-      const currentOverrides = Array.isArray(browserSearchSettings?.webSearchBangOverrides)
-        ? browserSearchSettings.webSearchBangOverrides
-        : [];
-      const nextOverrides = currentOverrides.filter((override) => override.key !== key);
-      if (changed && aliases.length > 0) {
-        nextOverrides.push({ key, aliases });
-      }
-      const sortedOverrides = nextOverrides.sort((a, b) => a.key.localeCompare(b.key));
-      await window.electron.saveSettings({
-        browserSearch: {
-          ...browserSearchSettings,
-          webSearchBangOverrides: sortedOverrides,
-        },
-      });
-      setWebSearchBangOverrides(sortedOverrides);
-    } catch (error) {
-      console.error('Failed to save web search bang aliases:', error);
-    } finally {
-      closeWebSearchBangPrompt();
-    }
-  }, [closeWebSearchBangPrompt, webSearchBangPrompt]);
-
-  const toggleWebSearchBangDisabled = useCallback(async (result: WebSearchResult | null) => {
-    if (!result?.bangKey) return;
-    const key = result.bangKey;
-    try {
-      const currentSettings = await window.electron.getSettings();
-      const browserSearchSettings = currentSettings.browserSearch;
-      const currentDisabled = Array.isArray(browserSearchSettings?.webSearchDisabledBangKeys)
-        ? browserSearchSettings.webSearchDisabledBangKeys
-        : [];
-      const disabledSet = new Set(currentDisabled);
-      if (disabledSet.has(key)) {
-        disabledSet.delete(key);
-      } else {
-        disabledSet.add(key);
-      }
-      const nextDisabled = Array.from(disabledSet).sort();
-      await window.electron.saveSettings({
-        browserSearch: {
-          ...browserSearchSettings,
-          webSearchDisabledBangKeys: nextDisabled,
-        },
-      });
-      setWebSearchDisabledBangKeys(nextDisabled);
-    } catch (error) {
-      console.error('Failed to update disabled bang:', error);
-    }
-  }, []);
-
-  const toggleWebSearchShowHidden = useCallback(async () => {
-    const next = !webSearchShowHiddenBangs;
-    setWebSearchShowHiddenBangs(next);
-    setWebSearchSelectedIndex(0);
-    try {
-      const currentSettings = await window.electron.getSettings();
-      await window.electron.saveSettings({
-        browserSearch: {
-          ...currentSettings.browserSearch,
-          webSearchShowHiddenBangs: next,
-        },
-      });
-    } catch (error) {
-      console.error('Failed to save hidden bang visibility:', error);
-    }
-  }, [webSearchShowHiddenBangs]);
-
-  const openWebSearchCustomBangPrompt = useCallback(() => {
-    setWebSearchCustomBangPrompt({
-      key: '',
-      aliases: '',
-      name: '',
-      host: '',
-      template: 'https://example.com/search?q={query}',
-    });
-  }, []);
-
-  const closeWebSearchCustomBangPrompt = useCallback(() => {
-    setWebSearchCustomBangPrompt(null);
-    window.setTimeout(() => webSearchInputRef.current?.focus(), 0);
-  }, []);
-
-  const saveWebSearchCustomBang = useCallback(async () => {
-    if (!webSearchCustomBangPrompt) return;
-    const key = normalizeWebSearchBangAliasList(webSearchCustomBangPrompt.key)[0] || '';
-    const aliases = normalizeWebSearchBangAliasList(webSearchCustomBangPrompt.aliases).filter((alias) => alias !== key);
-    const name = webSearchCustomBangPrompt.name.trim();
-    const host = webSearchCustomBangPrompt.host.trim().replace(/^https?:\/\//i, '').replace(/\/.*$/, '');
-    const template = webSearchCustomBangPrompt.template.trim().replace(/\{\{\{s\}\}\}/g, '{query}');
-    if (!key || !name || !host || !template.includes('{query}')) return;
-    try {
-      const currentSettings = await window.electron.getSettings();
-      const browserSearchSettings = currentSettings.browserSearch;
-      const currentProviders = Array.isArray(browserSearchSettings?.webSearchBangCustomProviders)
-        ? browserSearchSettings.webSearchBangCustomProviders
-        : [];
-      const nextProviders = [
-        ...currentProviders.filter((provider) => provider.key !== key),
-        { key, aliases, name, host, template },
-      ].sort((a, b) => a.key.localeCompare(b.key));
-      await window.electron.saveSettings({
-        browserSearch: {
-          ...browserSearchSettings,
-          webSearchBangCustomProviders: nextProviders,
-        },
-      });
-      setWebSearchBangCustomProviders(nextProviders);
-      closeWebSearchCustomBangPrompt();
-    } catch (error) {
-      console.error('Failed to save custom bang:', error);
-    }
-  }, [closeWebSearchCustomBangPrompt, webSearchCustomBangPrompt]);
-
-  useEffect(() => {
-    if (!webSearchBangPrompt) return;
-    const onKeyDown = (event: KeyboardEvent) => {
-      if (event.key === 'Escape') {
-        event.preventDefault();
-        closeWebSearchBangPrompt();
-        return;
-      }
-      if (
-        (event.key === 'Enter' || event.code === 'Enter' || event.code === 'NumpadEnter') &&
-        !event.altKey &&
-        !event.ctrlKey &&
-        !event.metaKey &&
-        !event.shiftKey
-      ) {
-        event.preventDefault();
-        void saveWebSearchBangAliases();
-      }
-    };
-    window.addEventListener('keydown', onKeyDown, true);
-    return () => window.removeEventListener('keydown', onKeyDown, true);
-  }, [closeWebSearchBangPrompt, saveWebSearchBangAliases, webSearchBangPrompt]);
-
-  useEffect(() => {
-    if (!webSearchCustomBangPrompt) return;
-    const onKeyDown = (event: KeyboardEvent) => {
-      if (event.key === 'Escape') {
-        event.preventDefault();
-        closeWebSearchCustomBangPrompt();
-        return;
-      }
-      if (
-        (event.key === 'Enter' || event.code === 'Enter' || event.code === 'NumpadEnter') &&
-        event.metaKey &&
-        !event.altKey &&
-        !event.ctrlKey
-      ) {
-        event.preventDefault();
-        void saveWebSearchCustomBang();
-      }
-    };
-    window.addEventListener('keydown', onKeyDown, true);
-    return () => window.removeEventListener('keydown', onKeyDown, true);
-  }, [closeWebSearchCustomBangPrompt, saveWebSearchCustomBang, webSearchCustomBangPrompt]);
-
-  useEffect(() => {
-    if (webSearchQuery === null || webSearchBangPrompt) return;
-    const onKeyDown = (event: KeyboardEvent) => {
-      if (event.metaKey && !event.ctrlKey && !event.altKey && !event.shiftKey && (event.key === 'n' || event.key === 'N')) {
-        event.preventDefault();
-        event.stopPropagation();
-        if (selectedWebSearchResult?.kind === 'bang') {
-          openWebSearchBangPrompt(selectedWebSearchResult);
-        } else {
-          openWebSearchCustomBangPrompt();
-        }
-        return;
-      }
-      if (
-        event.metaKey &&
-        !event.ctrlKey &&
-        !event.altKey &&
-        !event.shiftKey &&
-        (event.key === 'd' || event.key === 'D') &&
-        selectedWebSearchResult?.kind === 'bang'
-      ) {
-        event.preventDefault();
-        event.stopPropagation();
-        void toggleWebSearchBangDisabled(selectedWebSearchResult);
-      }
-    };
-    window.addEventListener('keydown', onKeyDown, true);
-    return () => window.removeEventListener('keydown', onKeyDown, true);
-  }, [openWebSearchBangPrompt, openWebSearchCustomBangPrompt, selectedWebSearchResult, toggleWebSearchBangDisabled, webSearchBangPrompt, webSearchQuery]);
+  submitBrowserSearchRef.current = submitBrowserSearch;
 
   const { runLocalSystemCommand } = useLauncherLocalSystemCommands({
     expandLauncherForDirectLaunch,
@@ -3016,11 +2425,6 @@ const App: React.FC = () => {
 
   // ─── Web Search mode ─────────────────────────────────────────────
   if (webSearchQuery !== null) {
-    const isWebSearchBangManager = !String(webSearchQuery || '').trim() ||
-      webSearchBangState.mode === 'selecting' ||
-      webSearchShowHiddenBangs;
-    const activeWebSearchBang = webSearchBangState.mode === 'active' ? webSearchBangState.bang : null;
-
     return (
       <WebSearchView
         alwaysMountedRunners={alwaysMountedRunners}
@@ -3039,11 +2443,7 @@ const App: React.FC = () => {
         selectedResult={selectedWebSearchResult}
         activateResult={activateWebSearchResult}
         submitSearch={submitBrowserSearch}
-        loadMoreResults={() =>
-          setWebSearchVisibleResultCount((count) =>
-            Math.min(webSearchResults.length, count + WEB_SEARCH_VISIBLE_RESULTS_INCREMENT)
-          )
-        }
+        loadMoreResults={loadMoreWebSearchResults}
         effectiveSearchBangs={effectiveSearchBangs}
         activeBang={activeWebSearchBang}
         isBangManager={isWebSearchBangManager}
